Update project settings and dependencies
This adds a Docker container to run locally for development and for
testing, as well as updates various dependencies to latest versions,
and moves to CircleCI for testing and publishing.

This commit differs from the prior commits by using sbt 1.0.4 which
fixes a problem with `sbt publish`.
See: https://github.com/sbt/librarymanagement/pull/183